refactor: 콘솔 환경에서 사용되는 InputView, OutputView 파일을 `src/view/console/` 로 이동
